@@ -2,29 +2,29 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20101112111650) do
-
-  create_table "boards", :force => true do |t|
-    t.string   "title"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+ActiveRecord::Schema[7.1].define(version: 2025_09_29_184342) do
+  create_table "boards", force: :cascade do |t|
+    t.string "title"
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
   end
 
-  create_table "posts", :force => true do |t|
-    t.string   "svg"
-    t.string   "title"
-    t.string   "description"
-    t.integer  "board_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+  create_table "posts", force: :cascade do |t|
+    t.string "svg"
+    t.string "title"
+    t.string "description"
+    t.integer "board_id"
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
+    t.index ["board_id"], name: "index_posts_on_board_id"
   end
 
 end
